Correction affichage trajectoire

+++ interfaces/TrajectoriesGenerator.java
@@ -6,5 +6,5 @@ import classes.Trajectory;
 
 public interface TrajectoriesGenerator {
     /** Returns a list of trajectories computed from the current state of the system. */
-    public List<Trajectory> generate(int steps);
+    public List<Trajectory> generate(double steps);
 }
